Complete ability to merge multicurrencies into one
merge Payee and Memo from other currency into ours, including currency
conversion. In this system, the foreign currency is completely lost, and
transactions appeare as if they were in our local currency.

The process here is to find the payee and memo name from the foreign currency
transaction and copy it in the currency conversion transaction in our currency.